removing unused imports and code lines
