+++ src/data-mgt/python/devices-tranformation/utils.py
@@ -1,11 +1,26 @@
+from datetime import datetime
+
 import pandas as pd
+
+
+def str_to_date(st):
+    """
+    Converts a string to datetime
+    """
+    return datetime.strptime(st, '%Y-%m-%dT%H:%M:%S.%fZ')
+
+
+def date_to_str(date):
+    """
+    Converts datetime to a string
+    """
+    return datetime.strftime(date, '%Y-%m-%dT%H:%M:%S.%fZ')
 
 
 def handle_api_error(api_request):
     json = None
 
     try:
-        # Try to parse json and check if body contains a specific error message.
         json = api_request.json()
         print(api_request.request.url)
         print(api_request.request.body)
